Refactor form folder (at least native type components)
