+++ react/components/EditorContainer/Sidebar/ComponentEditor/Form.tsx
@@ -4,10 +4,10 @@ import JsonSchemaForm, { FormProps, Widget, WidgetProps } from 'react-jsonschema
 
 import ArrayFieldTemplate from '../../../form/ArrayFieldTemplate'
 import BaseInput from '../../../form/BaseInput'
-import BrandSelector from '../../../form/BrandSelector/index'
-import CategorySelector from '../../../form/CategorySelector/index'
-import CollectionSelector from '../../../form/CollectionSelector/index'
-import DepartmentSelector from '../../../form/DepartmentSelector/index'
+import BrandSelector from '../../../form/components/BrandSelector/index'
+import CategorySelector from '../../../form/components/CategorySelector/index'
+import CollectionSelector from '../../../form/components/CollectionSelector/index'
+import DepartmentSelector from '../../../form/components/DepartmentSelector/index'
 import Dropdown from '../../../form/Dropdown'
 import ErrorListTemplate from '../../../form/ErrorListTemplate'
 import FieldTemplate from '../../../form/FieldTemplate'
