+++ react/components/form/components/CategorySelector/queries/Categories.graphql
