+++ react/components/form/components/CollectionSelector/queries/Collections.graphql
