+++ react/components/form/components/DepartmentSelector/queries/Departments.graphql
